Strip down test conn.log.
zeek-6.0 will change what is considered a local net and what is not.
This causes changes to `conn.log`, so we take this opportunity to only
check needed information in the log. This should lead to less random
test breakage in the future should Zeek ever change what is in this log
again.