Fix computation of message length.
`ResourceRecord.rdlen` is the length of the payload in `bytes`, but we
used it as the number of entries for `TXT` records. With this patch we
compute the correct length by setting up a view of the correct size;
this also fix issues with different record types where they would not
consume not the correct amount of input.

Closes #10.
Closes #11.